third_party/blink/web_tests/virtual/async-script-scheduling-apply-to-allowlist/wpt_internal/async-script-scheduling/execution-order.sub_default-expected.txt

This is a testharness.js-based test.
[FAIL] Async Script Execution Order
  Uncaught Error: assert_equals: expected "async_preload_or_blocking_render;async_preload_or_blocking_render;async;async;async;inline1;external;inline2;defer;DOMContentLoaded" but got "async_preload_or_blocking_render;async_preload_or_blocking_render;async;async;inline1;external;inline2;defer;DOMContentLoaded;async"
Harness: the test ran to completion.
